Post on games, starter logic for auto wanted_game post
Post is made on games when an api-game is clicked on followed by a
click submission on the add button.
On success the response is returned as data and a post to wanted_games
is intiated as a callback, using the game_id returned from createGameSuccess
as a callback.
As of right now the this attempt is returning a 400 error 'bad request'
so I must not be giving the data in a way that the api expects it.