UX and UI improvements
Index GET request now occurs when document readies.
Wanted games table now has proper headings.
Api games table now shows the date in a readable format as opposed to
the epoch format that it was previously in.
Tables are no longer visible prior to signing in.
Tables are hidden when user signs out.