Modularized some code and added more fields to games
The class activation logic was used for each dataTable so I turned it
into a function and invoked the function in place of the repeated code.

The games table now has several more columns so code was needed to access
that data and use it when creating a new game.