Improved Ux
If a game does not have a storyline or summary, a message will indicate
such in handlebars instead of just leaving the space blank.
Added toggleTables function in ui.js to toggle the visibility of the
tables when signing in and out.
Added functions in events.js to hide and show the tables when a game in
double clicked on, and the back button is hit from the single game view
respectively.
Added buttons to make a Post to games and wanted_games.
Added delete button to the show view of wanted_games to delete a
wanted_game.