Error handling and console.log removal
Removed the majority of console.logs.
Created a function to check if the content div is empty, to determine
where to send the error message to the user, when they try to add a game
to the wishlist that already exists.
Added new classes to h4 message tags for individual targeting.